raa489000: Sink VBUS before switching to TCPC side
When the RAA48900 switches from charger IC control to TCPC control, it
will disable the ASGATE.  In cases where the board is booted without a
battery, this can cause a loss in VSYS.  This commit has the board
attempt to sink VBUS if it's present prior to switching the part to
TCPC control.

BUG=b:181712325
BRANCH=dedede
TEST=Build and flash galith, remove battery, plug in 45W charger,
verify that DUT can boot up.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Id5b3269911d3b8222b28886f66ef724e3e82c637