[DNS/WIP] usbc: Send missed disconnect mode during dual role swap.
When the mux is in connected state and dual role swap happens,
mux should be first set to disconnected state and then to connected
state with the new data role.
Change is implemented to set the disconnect latch flag, which inturn
sends disconnect mode to kernel before updating the kernel with the
new data role.

BUG=b:182826116
BRANCH=none
TEST=verified the changes by sending 'cc off' followed by 'cc on'
     commands using servo v4 uart port. As typec charger is connected
     to the servo v4, initially port role set to 'SNK UFP' and
     later changes to 'SNK DFP'.'cc off' and 'cc on' commands successfully
     disconnects and connects devices behind servo v4.

Change-Id: I809b4c6690931e409cc7b25bb5e1246618618c5e
Signed-off-by: Deepti Deshatty <deepti.deshatty@intel.corp-partner.google.com>