zephyr: hooks: store deferred data in .rodata on non-posix
Change the deferred data section from .data.hooks to .rodata.hooks,
matching the ECOS behavior. This was apparently done for compatibility
with native_posix, so only make this change when we are not on
ARCH_POSIX.

The change means that the hooks data is not copied in RAM anymore, which
saves a significant amount of device memory.

For volteer, before:
   text    data     bss     dec     hex filename
 223934    9716   40092  273742   42d4e build-ro/zephyr/zephyr.elf

After:
   text    data     bss     dec     hex filename
 225730    7924   40092  273746   42d52 build-ro/zephyr/zephyr.elf

Compared to ECOS:
   text    data     bss     dec     hex filename
 203680    6384   33676  243740   3b81c volteer/RO/ec.RO.elf

BUG=b:183748844
BRANCH=none
TEST=build and run on volteer

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ia23e9c0dc5ee9eaeee674f943b22762355ad5efe