zephyr: zmake: enable verbose logging with --debug
The --debug parameter (which is passed in the CQ, since we want to see
as much failure context as we can in the logs) currently prevents
disabling tracebacks.

We had some discussion when that landed about making it enable verbose
logging as well.  Make the default with --debug as DEBUG instead of
WARNING.

BUG=none
BRANCH=none
TEST=zmake -D testall => observed log level is DEBUG
     zmake testall => observed log level is WARNING
     zmake -l INFO testall => observed log level is INFO

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I87bf1d7cca34cffd9ff6deeb53fecf0ae8f2e362