TCPMv2: Set CC on contract invalidation only if connected
When a contract is invalidated, it's important we advertise our current
limit Rp rather than a residual collision avoidance Rp.  However, we
should not attempt to set CC's when we're in a disconnected state. This
may cause us to set an unexpected CC selection, such as Ra if the TC
layer hasn't set the first termination yet.

BRANCH=None
BUG=b:183182447
TEST=on guybrush, boot with i2c tracing enabled and ensure Ra is never
set during init.  Perform PR swaps and hard resets as source to ensure
correct Rp is set.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I8fbe06685e4b4a6493dfbf55270ce3bda8cfbdfd