zephyr: create a Kconfig option for CONFIG_CMD_SYSINFO
Create a Kconfig option for the sysinfo console command, furthering
our parity with EC configurability.

This is by default enabled anyway, so the Kconfig is just providing a
means to turn it off, if desired.  (Which is useful for ITE chips
right now, where we don't have full support for all the deps we bring
in with this command yet.)

BUG=b:180980668
BRANCH=none
TEST=hayato build (follow-up CL)

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ibe2c2a4d7620ef60979b256735046379bf99e3e5