zephyr: sensors: Add alternate sensors support
In some boards there's a runtime decision for which sensor to use. In
order to support this we'll need an instance of the alternate sensors
as well as a mapping for them.

Add an optional phandle `alternate-for` to the base sensors bindings.
When used, the sensor will be excluded from the priomary
motion_sensors array and added instead to the motion_sensors_alt
array.

BRANCH=none
BUG=b:183990188, b:185966444
TEST=zmake testall

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I6f6e3a9659a381c45fd97fbce9fecb2a807f7fa2