it83xx: watchdog: introduce timer 1 work as warning timer
Timer 1 (16bit) is suitable for working as the warning timer.
Because it has the same clock source with watchdog counter
(current setting is 1.024khz).
So the 16bit timer is enough to print out pre-warning messages
until reset.

With this CL, we free a 32bit timer for use.

BUG=b:186028167
BRANCH=none
TEST=enable CONFIG_IT83XX_WDT_WARNING_BY_TIMER1 and run
console command "waitms":
"waitms 1200", EC warning but no reset.
"waitms 1600", EC warning and reset.

Signed-off-by: Dino Li <Dino.Li@ite.com.tw>
Change-Id: I8b08047663a24cb1733aaa2f51be8cb4ab4f0fae