Battery: Add command to export display SoC
Currently, CrOS EC passes the battery remaining capacity (mAh) and the
full capacity (mAh) through ACPI to the AP so that the host can
calculate the battery SoC.

The host further manipulates the SoC to get the display SoC, which is
used to determine user visible behaviors. To get the consistent
behaviors in all power states, this change enables the EC to send the
display SoC to the host via EC_CMD_DISPLAY_SOC command.

BUG=b:174433637, b:181506409
BRANCH=dedede, trogdor
TEST=Storo, CoachZ

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Idc6992625d992a73be141987d02ed220508d3b74