usb_pe_drp: Consider CHARGE_PORT_NONE for HardRst
When we arrive in the EC image without a battery and with a PD
supplier, we would eventually consider sending a HardReset since we
are expecting SourceCaps to be sent by the port partner.  However,
since we no longer maintain our contract across images and we would
like for the board to not brown out, we avoid sending the HardReset
and simply disable the port instead.

However, this situation can occur before we decide on a charge port.
Therefore, this commit simply adds another condition when deciding to
skip sending the HardReset.  If a battery is not present, but there's
no active charge port, we must be running on AC somehow so don't send
the HardReset as this port may in fact be the previous charge port.

BUG=b:176214112,b:178728138
BRANCH=dedede
TEST=Build and flash galith; plug in charger on C1, remove battery,
sysjump between RO and RW many times, verify that SoC remains in S0.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I16054d9c1ddac780243524e23a656ba3770bb4a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2848290
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Auto-Submit: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Diana Z <dzigterman@chromium.org>