zephyr: shim: do not return from main if thread monitoring is enabled
Returning from main causes the corresponding thread structure to be
skipped from the thread list, so that it does not get displayed by
commands such as "kernel stacks". The structure is statically allocated
though, so it is convenient to keep it active when thread monitoring is
enabled so that the corresponding stack can be tuned.

After this patch:

uart:~$ kernel stacks
...
0x200c76f8 main (real size 1024): unused 148 usage 876 / 1024 (85 %)
...

BUG=b:183748844
BRANCH=none
TEST=kernel stacks

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Iee34fac36f5e8989e9d7c761b7cc6b9a14a7b987