zephyr: Add CBI SSFC support
Some boards support different motion sensors based on the CBI SSFC field
stored in EEPROM. The decision about which sensor is made in runtime
thus all drivers have to be built-in.

Define structure of the SSFC in the devive tree("named-cbi-ssfc"),
that allows using generic driver instead of board-specific code as it
is done in CrosEC.

Every SSFC field value("named-cbi-ssfc-value") is associated with
an alternative sensor(or sensors) which will be used (instead of the one
pointed by 'alternative-for' property) if the value in
EEPROM matches.

BUG=b:183990188
BRANCH=none
TEST=Add alternative motion sensors to the device tree, modify CBI SSFC
with 'cbi set 8 value 4', reboot EC and verify that the new sensors are
used with the 'accelinfo' command.

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I3b5f3c171005885d96b1fdf14e844aaf862b6818