isl9241: fix the system can't boot due to DCM mode
When the device recover from the battery cutoff,
the battery is not ready to output the power to the
system. If we toggle the DCM mode at that time, the
system power will be cutoff. To avoid this problem,
we check the battery battery_get_disconnect_state()
before setting the DCM.

BUG=b:186188004
BRANCH=firmware-volteer-13672.B
TEST=1. charge the battery to near full
     2. battery cutoff
     3. plug in AC

Change-Id: Ic2d959c89a9e37479b919133e10a69be8148a26e
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>