zephyr: Kconfig: disable CONFIG_TIMESLICING
This adds a Kconfig file for overriding Zephyr defaults and adds an
entry to disable CONFIG_TIMESLICING. The feature is compiled in by
default but not enabled, and we don't enable it at runtime. Compiling it
out saves a few bytes.

Before:
FLASH:      233760 B       512 KB     44.59%
 SRAM:       48344 B        62 KB     76.15%

After:
FLASH:      233344 B       512 KB     44.51%
 SRAM:       48328 B        62 KB     76.12%

BUG=b:183748844
BRANCH=none
TEST=build and run on volteer

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I5587a553197ca0101e9d7047f6ce7a988dd0b36d