honeybuns: Fix discover mode reply for USB 2.0 signals
Bit 7 of the DP capabilities (Discover mode response) message is for
USB r2.0 Signaling Not Used. This bit was set for honeybuns, which is
incorrect. This can cause some host machines (Mac) to stop enumerating
USB r2.0 devices when 4 lane DP is selected.

BUG=b:186629871
BRANCH=quiche
TEST=Provided ODM partner with this bit being set to 0 and they
verifed that USB2.0 enumeration is not lost when DP is connected.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I137a41dd641c0c019a649c5fbcc37bc44775e30a