tcpc:  Add common routine to check power
This comment is from CL:1123314:
We do not want to allow depthcharge to suspend the PD port for a
firmware upgrade if there is not enough battery power to last during
the full firmware upgrade.

A common routine is created to check power:
pd_firmware_upgrade_check_power_readiness() in usb_common.c.

Firmware upgrade which needs to suspend PD port should check
this function to make sure there is enough power.

BUG=none
BRANCH=none
TEST=make -j buildall

Signed-off-by: li feng <li1.feng@intel.com>
Change-Id: I47910a4d82da9b569c297ad2cd399a2faf4cf98a