tcpc: call common function to check power
We do not want to allow depthcharge to suspend the PD port for a
firmware upgrade if there is not enough battery power to last during
the full firmware upgrade.

The check has moved to common function:
pd_firmware_upgrade_check_power_readiness() in usb_common.c. Calling
this function instead.

BUG=none
BRANCH=none
TEST=make -j buildall

Signed-off-by: li feng <li1.feng@intel.com>
Change-Id: I47910a4d82da9b569c297ad2cd399a2faf4cf98a