zephyr: fix namespace for Kconfig check scripts
Update both the scripts used to verify ad-hoc CONFIG options have a
Kconfig option to use the same search criteria.

The list of Kconfigs is now all option that start with the config or
menuconfig keyword. The full Kconfig list is transformed to replace
CONFIG_PLATFORM_EC_ with CONFIG_, to match the Chromium EC namespace.

BUG=b:184037201
BRANCH=none
TEST=make buildall
TEST=zmake testall

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I5edeeddd4e935b4bbf2ce7321a19249a543d916e