util: Convert CONFIG checker to use kconfiglib
Update this to use kconfiglib so that it might be easier to maintain.
This needs a more correct Kconfig setup in the tests. Also we need to
provide the root Kconfig instead of just the directory.

Overall this takes about 50% more time, but the total is only about
50ms, so this is not a significant factor in build performance.

BUG=b:181323955
BRANCH=none
TEST=python3 util/test_kconfig_check.py

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I5cce2ca1fa16a7c334cef15b1f66dd7ede2644bd