Guybrush: Correct temperature indexes
The temp_sensor index should indicate the "index among the same kind of
sensors."  In the case of thermistors, this will be directly used as the
ADC index to read from.

Note that this is a no-op for guybrush, because the adc_channel indexes
already aligned with the temp_sensors indexes.

BRANCH=None
BUG=None
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ia316a8985271cfc2af070ad351a0aba94703dc2a