Makefile: Update C++ compile command
Add cmd_cxx_to_host, which is the C++ equivalent of cmd_c_to_host. Use
clang++ to match "CXX" in Makefile.toolchain (and the toolchain that we
use in the rest of Chrome OS). Add "-std=gnu++17" to match the C++
standard that we use in the rest of Chrome OS.

Followup commits will make use of these changes.

BRANCH=none
BUG=b:144959033
TEST=make buildall

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I5878ae64dc5ac76db22b3c331915b191a00f4257