NX20P348x: Only check sink bit for sink_enable error
When a port is sourcing (ex. to a dongle), running sink disable on the
port will currently return failure because the 5VSRC bit is set.
However, sinking has been successfully disabled.  Reflect this in the
error return by only checking the specific sinking bit in the status
register.

BRANCH=None
BUG=b:187220141
TEST=on guybrush, plug and unplug AC on C0 with a dongle on C1.  Verify
no failures to disable sinking are present.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I529d33b41dc4bc55f7c647742c70832a125fd367