TCPMv2: Re-enable SOP' transmission on interruption to Vconn swap
If a Vconn swap is interrupted before we receive a PS_RDY from the
partner, re-enable SOP' transition.  Note this doesn't matter for the
case of no PS_RDY received as we'll proceed to hard reset which will
also reset our Vconn role.  This is primarily applicable if the
interruption is for an unexpected message, which interrupts the swap and
leaves us the Vconn source.

BRANCH=volteer
BUG=b:186886218
TEST=on drobit, ensure cable entry doesn't fail after failed Vconn swap
with dock

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I952c04535ae9527f07cb203c403a51182c67cba7