asurada: drop SVY682X smart discharge.
As we have ADCs for VBUS detection, we don't need to
rely on the VBUS detection in SYV682X, so the VBUS SDSG
function is not needed anymore (to make VBUS fast falling in to Vsafe0V).

Also, SDSG might make concerns of FRS, since in transisiton, there
should be no discharge circuit happen or the swap might fail.

BUG=b:181203590
TEST=FRS 5V/9V/20V works on Asurada. Meets tVBUSOff.
BRANCH=asurada

Change-Id: I0d8f2a4b038b1ecdf23e834e37d59da9ea5d2c57
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>