virtual mux: Wait for mux ACK from AP on exiting alt mode
With an AC charger, EC enters USB mode on hotplug and sends a mux
diconnect after selecting the sink capability. With the current PD
flow, when we send a disconnect, PD waits for an ack plus 12.5msec.
This might cause tSenderResponse timer to expire and the source will
end up sending hard reset.

So, instead of waiting for an ack for every disconnect, only wait for an
ACK from AP when exiting an alternate mode.

BUG=b:187796705
BRANCH=None
TEST=1. Tested on Brya, connecting AC charger no longer cases hard
        resets
     2. Both the monitors behind the Tortilla based dock enumerate after
        coldboot.
     3. Checked Disconnect-Connect with TBT/USB4 device in S0ix, they
        enumerate fine on full resume.
     4. Checked coldboot with non-Tortilla based dock based dock. it
        enumerates fine.

Signed-off-by: Ayushee Shah <ayushee.shah@intel.com>
Change-Id: Ia408b45bbd6805e9f51191dc80f333839b2252ba