Mancomb: Add BJ to charge manager and correct active charge port set
Register the BJ with the charge manager, and add a per-board callback to
determine the voltage and current associated with it.

Additionally, correct a couple of errors in the active charge port set
function.  The BJ port is incapable of sourcing Vbus, and the ports to
be disabled should be all except the port being enabled.

BRANCH=None
BUG=b:187911213
TEST=on mancomb, plug in BJ and verify the correct voltage and current
are registered for it

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I25369e18a56832a79421e179d267c4741916720b