TCPMv2: Seed the charge manger for the reboot-SRC-PRS-SNK path
When EC reboots, the TCPMv2 stack initializes the Type-C supplier in
either Unattached.SNK or Attached.SNK, in order to seed the charge
manager.

However, if the initial state is not a SNK state, e.g. Attached.SRC,
the charge manger has never been seeded. It happens when a DRP (like
a Type-C hub) is attached during reboot. It transits to Attached.SRC
then perform a Power-Role-Swap, and transits to Attached.SNK.

In this Power-Role-Swap case, need to initialize the Type-C supplier
to 0 in order to seed the charge manage.

BRANCH=None
BUG=b:188532782
TEST=Attach a powered Type-C hub; reboot EC; the charge manger is
seeded; the charge port is selected; the power sink path is enabled.

Change-Id: I8fbfc1c4ffc5c9a9a8a49fcc36d67da93d01c403
Signed-off-by: Wai-Hong Tam <waihong@google.com>