brya: Add lid_open interrupt annotation
This updates the GPIOD2 (the lid open interrupt) pin alternate mode
declaration to include the interrupt declaration. This is done for
consistency by the gpio.inc generator; there is no functional change
as the GPIO_INT() declaration already sets the interrupt mode.

Note: The reason the existing declaration works is because the npcx
support code treats interrupt flags as additive w.r.t. the GPIO(...)
declaration. Other flags are not additive and must be specified a 2nd
time. Best practice is to always specify the complete set of config
flags.

BRANCH=none
BUG=b:185322560
TEST=visual inspection

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: I29c5ed44479dfcc732d01d46ccab5ed1ecaa314b