zephyr: npxc: Fix non-psl hibernate
The number of RAM blocks depends on the NPCX7 chip variant.
Define NPCX7_RAM_BLOCK_NUM based on the CONFIG_SOC_NPXC7x instead of
hardcoding it to 12.

This change should fix the hibernate process for NPCX7M6 chip variants.

BUG=b:188409672
BRANCH=none
TEST=Build and flash Zephyr on Lazor, hibernate the DUT(Alt+h+VolUp or
'hibernate' console command) and wake it up with a PB press.

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: Ia4901de31898455021a65de5ef3775872e574377