zmake: Allow filter tests to be run without pytest
Put these tests in a class and add a main program so we can run the
tests directly.

This makes it easier to see any output produced by the test, for
debugging purposes. For example a print() added to a test will not be
supressed. This makes debugging easier.

BUG=b:184298184
BRANCH=none
TEST=(cd zephyr/zmake/; PYTHONPATH=`pwd` python3 tests/test_zmake.py
	 TestFunctional.test_filter )

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ic5384d8926e54322a4bf828dd6ee91b1ab4507ff