USBC: Disallow Try.SRC when not toggling
Chromeboxes may Try.SRC when a dedicated charger is present.  However,
they should not Try.SRC at a time when they are note dual-role toggling.
This likely represents a time when they are unable to source Vbus and
should therefore not attempt to become the source.

BRANCH=None
BUG=b:188710940
TEST=on mancomb, power on with the BJ charger and then connect a dock in
G3.  Observe we do not attempt to source the dock sans Vbus

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ia9062f8f833d678b5b3c9daf576dc6ad2f2a8c74