brya: Skip mp2964 tuning on expected AP global reset
Due to CSE lite boot behavior, the AP goes through an additional global
reset. This shows up as a transition back down to S5 on the
EC. CHIPSET_SHUTDOWN and CHIPSET_STARTUP hooks are executed when this
happens, so we need to remember that we've been here before and only run
mp2964 update the first time through.

BRANCH=none
BUG=b:185424011
TEST=verified mp2964 tuning only runs once using EC console logs

Change-Id: I08007dfa62bf842c51e4f6474322302409f4256c
Signed-off-by: Caveh Jalali <caveh@chromium.org>