brya: Board ID 1: Handle TCPC_RST reassignment
We are transitioning GPIO definitions to be correct for board ID 2. In
order to support board ID 1 with the same EC image, some GPIOs need to
be reconfigured to their legacy settings at runtime when board ID 1 is
detected. For board ID 2, the TCPC C0/C2 reset GPIO has moved to a
previously unused pin. The original pin is now an ADC input pin.

BRANCH=none
BUG=b:183452273
TEST=verified TCPC C0/C2 can be reset on board ID 1 using
	ID_1_USB_C0_C2_TCPC_RST_ODL.

Change-Id: I52d8044ed10379346ae36d4f5d6cbe7446867182
Signed-off-by: Caveh Jalali <caveh@chromium.org>