brya: Board ID 1: Handle KB_BL_EN reassignment
We are transitioning GPIO definitions to be correct for board ID 2. In
order to support board ID 1 with the same EC image, some GPIOs need to
be reconfigured to their legacy settings at runtime when board ID 1 is
detected. For board ID 2, the keyboard backlight enable GPIO has moved
to a previously unused pin and its polarity is inverted. The original
pin is now an ADC input pin.

BRANCH=none
BUG=b:183452273
TEST=verified keyboard backlight function with "kblight" EC console command

Change-Id: I86a1b09c9aaab8f6275a65cd1331f135b152f538
Signed-off-by: Caveh Jalali <caveh@chromium.org>