brya: Board ID 1: Handle BATT_PRES reassignment
We are transitioning GPIO definitions to be correct for board ID 2. In
order to support board ID 1 with the same EC image, some GPIOs need to
be reconfigured to their legacy settings at runtime when board ID 1 is
detected. For board ID 2, the battery presence detect GPIO has moved to
the previous keyboard backlight enable pin.

BRANCH=none
BUG=b:183452273
TEST=booted on brya board ID 1, check battery present status on
	ID_1_EC_BATT_PRES_ODL.

Cq-Depend: chromium:2914207
Change-Id: Ibae2a5e1f43c83360535e0d60c2a343bf9ef2421
Signed-off-by: Caveh Jalali <caveh@chromium.org>