usb_port_power_dumb: allow board to customize hooks
On Asurada, some Sandisk stickers would enter fault status
(down-train to USB2 or not detected) if Vbus is applied before xhci
initialization. The defauilt S3 hooks does not work in this case.

This CL adds a config option to allow board to control the timing to
enable/disable port power.

BUG=b:187149602
TEST=manually test with CL:2909972
BRANCH=asurada

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: I2ac4689182f22f4fa81d34bff8b5797e5fe222d2