stm32f4: i2c: send slave response faster
Since TxE is not set in the address phase, we should start responding
right after we have something in the buffer, as we might end up not
receiving more interrupts.
EC, however, must be treated differently as the result might not be
ready yet.

BRANCH=none
BUG=none
TEST=`make BOARD=rose -j`, getting status or hdesc should not result in
locking of SCL line (should not be stretched indefinetely).

Change-Id: I9c7a898738b7eeb0f4edd65f1f9a504a97d4ae94
Signed-off-by: Angela Czubak <acz@semihalf.com>