tpm_test: drbg_test: explicitly state whether to check the result
We won't have the responses for the lab vectors. Add a check_result bool
to the DRBG_GENERATE test items, so there's a way to tell when it needs
to record the result for the generate command. The test will compare
the expected result to the actual result while running if check_result
is True and the expected string  is not empty.

BUG=b:189376694
TEST=./tpmtest

Change-Id: Ifbb5ad9737a6649affce4c885083b631f5056a26
Signed-off-by: Mary Ruthven <mruthven@chromium.org>