Makefile: Switch to the Python script for CONFIG checking
Now that the Python script seems to do what we need, switch over to
use that for checking for allowed ad-hoc CONFIGs.

Sadly we need a work-around for the very old Python 3 version used in
the chroot.

The new script is better at finding Kconfig options, so this allows some
reductions in the config_allowed.txt file.

Delete the now-unused shell scripts.

BUG=b:181323955
BRANCH=none
TEST=python3 util/test_kconfig_check.py

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I2dafc9dfe9d9020638f4e49b5c5ee0fc0b10000b