chgstv2: Sustain battery SoC
This patch makes chgstv2 capable of sustaining the battery soc
(a.k.a. sustain soc). The host can set the lower and upper thresholds
between which the EC tries to keep the battery state of charge.

The sustain soc is off on reset. It's enabled or disabled by the host
using EC_CMD_CHARGE_CONTROL. For example, the host can disable it
on every resume and re-enable it on every suspend.

BUG=b:188457962
BRANCH=None
TEST=atlas

Change-Id: I62b4e8bc9517900a5a32d2f35369c645fa8a60c3
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>