honeybuns: Limit data role swap request attempts
This CL modifies honeybuns data role swap policy in 2 ways. First, the
dependence on dr_swap_flag is removed. Second, a new variable is added
to track data role swap attempts per each usbc attach. Previosly,
dr_swap_flag was used as a means to limit the number of data role swap
attempts, but the flag that is passed in is set in PE code that is
specific to a typical chromebook port policy.

The new policy is to always try to be the desired data role, but limit
the attempts for situations where the port partner rejects the
request. This case happens for example with compliance testing.

HOOK_USB_PD_CONNECT is used to clear the max attempt counter as this
should apply once per each usbc attach event.

BUG=b:188756664
BRANCH=quiche
TEST=Verifed that the host port will request a data role swap when
testing against the iPad. Against the ipad, host port always attaches
as a source, and would remain a DFP.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ib913f8bbba76d2bd12dccb6a6a49ef4249d428ed